# src/main/resources/contract/openApi.yaml
openapi: 3.0.0

info:
  description: >-
    Сервис предназначен для управления доставкой заказов и курьерами
  version: 0.0.1
  title: Delivery Api
servers:
  - url: 'http://localhost:8095/delivery/api/v1'
    description: Сервер управления доставкой
tags:
  - name: Courier
    description: API для управления курьерами
  - name: Order
    description: API для управления заказами
paths:
  /api/v1/orders:
    post:
      tags:
        - Order
      summary: Создать заказ
      description: Позволяет создать заказ с целью тестирования
      operationId: createOrder
      responses:
        '201':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderView'
        default:
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/orders/active:
    get:
      tags:
        - Order
      summary: Получить все незавершенные заказы
      description: Позволяет получить все незавершенные заказы
      operationId: getOrders
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderView'
        default:
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/couriers:
    post:
      tags:
        - Courier
      summary: Добавить курьера
      description: Позволяет добавить курьера
      operationId: createCourier
      requestBody:
        required: true
        description: Курьер
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourierCreateDto'
      responses:
        '201':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourierView'
        default:
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Ошибка выполнения бизнес логики
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags:
        - Courier
      summary: Получить всех курьеров
      description: Позволяет получить всех курьеров
      operationId: getCouriers
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourierView'
        default:
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    LocationView:
      type: object
      required:
        - x
        - y
      properties:
        x:
          type: integer
          description: X
          minimum: 0  # Валидация минимального значения для x
        y:
          type: integer
          description: Y
          minimum: 0  # Валидация минимального значения для y
    OrderView:
      type: object
      required:
        - id
        - location
      properties:
        id:
          type: string
          format: uuid
          description: Идентификатор
        location:
          $ref: '#/components/schemas/LocationView'
          description: Геолокация
    CourierCreateDto:
      type: object
      required:
        - name
        - speed
      properties:
        name:
          type: string
          description: Имя
          minLength: 1  # Валидация на минимальную длину
        speed:
          type: integer
          description: Скорость
          minimum: 1  # Валидация на минимальное значение
    CourierView:
      type: object
      required:
        - id
        - name
        - location
      properties:
        id:
          type: string
          format: uuid
          description: Идентификатор
        name:
          type: string
          description: Имя
        location:
          $ref: '#/components/schemas/LocationView'
          description: Геолокация
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: Код ошибки
        message:
          type: string
          description: Текст ошибки